pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "security-dataset-scraper"
version = "1.0.0"
description = "Comprehensive security dataset scraper for AI fine-tuning"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Security Dataset Team"}
]
keywords = ["security", "scraper", "dataset", "ai", "fine-tuning", "redteam"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.1",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "selectolax>=0.3.17",
    "html5lib>=1.1",
    "playwright>=1.40.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "tenacity>=8.2.0",
    "pyyaml>=6.0.1",
    "tiktoken>=0.5.0",
    "datasketch>=1.6.0",
    "aiolimiter>=1.1.0",
    "diskcache>=5.6.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "types-beautifulsoup4>=4.12.0",
    "types-aiofiles>=23.2.0",
    "types-PyYAML>=6.0.0",
]
ai = [
    "openai>=1.6.0",
    "anthropic>=0.8.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

[project.scripts]
security-scraper = "main:main"

[project.urls]
Homepage = "https://github.com/security-dataset/scraper"
Documentation = "https://security-dataset.github.io/scraper"
Repository = "https://github.com/security-dataset/scraper"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | data
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
skip = [".git", ".venv", "venv", "data"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "tests/",
    "data/",
    "build/",
    "dist/",
]

[[tool.mypy.overrides]]
module = [
    "selectolax.*",
    "datasketch.*",
    "aiolimiter.*",
    "diskcache.*",
    "playwright.*",
    "tiktoken.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
